requirements.txt
PyQt5
numpy
Pillow
pandas
matplotlib
scipy
opencv-python
